+++ src/pages/cookie-record/components/calendar-section/cookie-calendar/index.ts
@@ -1,0 +1,1 @@
+export { CookieCalendar, CalendarSkeleton } from './CookieCalendar'
