+++ src/pages/cookie-record/components/calendar-section/index.ts
@@ -1,0 +1,2 @@
+export { CalendarSection } from './CalendarSection'
+export { CalendarSkeleton } from './cookie-calendar'
