+++ src/pages/cookie-record/components/calendar-section/monthly-cookie-log-list/index.ts
@@ -1,0 +1,1 @@
+export { MonthlyCookieLogList } from './MonthlyCookieLogList'
