+++ src/pages/cookie-record/components/error-fallback/index.ts
@@ -1,0 +1,1 @@
+export { CookieRecordErrorFallback } from './ErrorFallback'
