+++ src/pages/cookie-record/components/header/index.ts
@@ -1,0 +1,1 @@
+export { CookieRecordHeader } from './Header'
