+++ src/pages/cookie-record/components/hint-drawer/answer-detail/AnswerDetail.tsx
