+++ src/pages/cookie-record/components/hint-drawer/answer-detail/index.ts
@@ -1,0 +1,1 @@
+export { AnswerDetail } from './AnswerDetail'
