+++ src/pages/cookie-record/components/hint-drawer/buy-hint-modal/BuyHintModal.tsx
