+++ src/pages/cookie-record/components/hint-drawer/buy-hint-modal/index.ts
@@ -1,0 +1,1 @@
+export { BuyHintModal } from './BuyHintModal'
