+++ src/pages/cookie-record/components/hint-drawer/error-fallback/error-modal/index.ts
@@ -1,0 +1,1 @@
+export { BuyHintErrorModal } from './ErrorModal'
