+++ src/pages/cookie-record/components/hint-drawer/error-fallback/index.ts
@@ -1,0 +1,1 @@
+export { BuyHintErrorFallback } from './ErrorFallback'
