+++ src/pages/cookie-record/components/hint-drawer/hint-list/index.ts
@@ -1,0 +1,1 @@
+export { HintList } from './HintList'
