+++ src/pages/cookie-record/components/hint-drawer/index.ts
@@ -1,0 +1,1 @@
+export { HintDrawer } from './HintDrawer'
