+++ src/pages/cookie-record/components/index.ts
@@ -1,0 +1,5 @@
+export * from './calendar-section'
+export * from './error-fallback'
+export * from './header'
+export * from './hint-drawer'
+export * from './log-section'
