+++ src/pages/cookie-record/components/log-section/cookie-log-list/cookie-log-text/hint-icon/index.ts
@@ -1,0 +1,1 @@
+export { HintIcon } from './HintIcon'
