+++ src/pages/cookie-record/components/log-section/cookie-log-list/index.ts
@@ -1,0 +1,1 @@
+export { CookieLogList } from './CookieLogList'
