+++ src/pages/cookie-record/components/log-section/index.ts
@@ -1,0 +1,1 @@
+export { LogSection } from './LogSection'
