+++ src/pages/cookie-record/hooks/index.ts
@@ -1,0 +1,4 @@
+export {
+  useCookieLogSearchParams,
+  cookieLogTypes,
+} from './useCookieLogSearchParams'
