+++ src/pages/cookie-record/index.ts
@@ -1,0 +1,1 @@
+export { default as CookieRecordPage } from './CookieRecordPage'
