+++ src/pages/index.ts
@@ -5,3 +5,4 @@ export * from './auth'
 export * from './profile'
 export * from './main'
 export * from './invite'
+export * from './cookie-record'
